+++ packages/utils-connector-tools/src/normalize-headers.ts
@@ -1,9 +1,10 @@
 import { HttpHeaders } from 'hint/src/lib/types';
 
+/** Normalize all keys of an `HttpHeader` to lowercase. */
 export const normalizeHeaders = (headers?: HttpHeaders | null) => {
     if (headers) {
         return Object.keys(headers).reduce((result, key) => {
-            result[key.toLowerCase()] = headers[key];
+            result[key.toLowerCase().trim()] = headers[key];
 
             return result;
         }, {} as HttpHeaders);